moved App into a components directory
