+++ .gitignore
@@ -1,4 +1,3 @@
 
 client/node_modules/*
-
-
+client/coverage/*
