+++ client/src/index.js
@@ -1,9 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
+import App from './components/App/App';
 import './index.css';
 
 ReactDOM.render(
   <App />,
   document.getElementById('root')
 );
+
